requirements.txt
Django==4.2.11
djangorestframework==3.14.0
Pillow==10.0.0
requests==2.31.0